refactor(appearance): share canonical atlas planning across surface samplers
